+++ app/lib/shared/versions.dart
@@ -21,7 +21,7 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// Make sure that at least two versions are kept here as the next candidates
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
-  '2021.10.28', // The current [runtimeVersion].
+  '2021.11.01', // The current [runtimeVersion].
   '2021.10.26',
   '2021.10.12',
 ];
@@ -56,7 +56,7 @@ final String runtimeSdkVersion = '2.14.1';
 final String toolStableDartSdkVersion = '2.14.4';
 final String toolStableFlutterSdkVersion = '2.5.3';
 final String toolPreviewDartSdkVersion = '2.15.0-178.1.beta';
-final String toolPreviewFlutterSdkVersion = '2.7.0-3.0.pre';
+final String toolPreviewFlutterSdkVersion = '2.7.0-3.1.pre';
 
 // Value comes from package:pana.
 final String panaVersion = pana.packageVersion;